@@ -5,7 +5,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -21,10 +21,12 @@ dependencies:
   # Network
   dio: ^4.0.0
   connectivity: ^3.0.6
+  retrofit: 2.0.1
 
   # Data storage
   shared_preferences: ^2.0.6
   flutter_secure_storage: ^4.2.1
+  hive: 2.0.4
 
   # DI
   get_it: ^7.2.0
@@ -33,19 +35,36 @@ dependencies:
   collection: ^1.15.0
   logging: ^1.0.2
   logger: ^1.1.0
-  json_annotation: ^3.1.1
+  json_annotation: 4.1.0
+  path_provider: 2.0.5
+  pretty_dio_logger: 1.2.0-beta-1
 
   # UI
   flutter_svg: ^0.22.0
 
-  #Locolization
+  #Localization
   intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.0.0
-  json_serializable: ^3.5.1
+  integration_test:
+      sdk: flutter
+  test: ^1.14.4
+  #Codegen
+  ### https://pub.dev/packages/build_runner
+  build_runner: 2.1.2
+  ### https://pub.dev/packages/retrofit_generator
+  retrofit_generator: 2.0.1+1
+  ### https://pub.dev/packages/json_serializable
+  json_serializable: 5.0.1
+  ### https://pub.dev/packages/hive_generator
+  hive_generator: 1.1.1
+
+  flutter_lints: ^1.0.0
+
+  # Lint rules
+  pedantic: 1.11.0
 
 flutter:
   uses-material-design: true
@@ -54,12 +73,12 @@ flutter:
     - assets/svg/
 
 flutter_intl:
-  enabled: true # Required. Must be set to true to activate the package. Default: false
-  class_name: S # Optional. Sets the name for the generated localization class. Default: S
-  main_locale: en # Optional. Sets the main locale used for generating localization files. Provided value should consist of language code and optional script and country codes separated with underscore (e.g. 'en', 'en_GB', 'zh_Hans', 'zh_Hans_CN'). Default: en
-  arb_dir: lib/l10n # Optional. Sets the directory of your ARB resource files. Provided value should be a valid path on your system. Default: lib/l10n
-  output_dir: lib/generated # Optional. Sets the directory of generated localization files. Provided value should be a valid path on your system. Default: lib/generated
-  use_deferred_loading: false # Optional. Must be set to true to generate localization code that is loaded with deferred loading. Default: false
+  enabled: true
+  class_name: S
+  main_locale: en
+  arb_dir: lib/localization/l10n
+  output_dir: lib/localization/generated
+  use_deferred_loading: false
 
 Default: false
 
