@@ -1,7 +1,8 @@
 import 'dart:ui';
 
 class AppColors {
-  static const transparent = Color(0x00000000);
-  static const white = Color(0xffffffff);
-  static const black = Color(0xff000000);
+  AppColors._();
+  static const Color transparent = Color(0x00000000);
+  static const Color white = Color(0xffffffff);
+  static const Color black = Color(0xff000000);
 }
