@@ -1,3 +1,6 @@
+import 'dart:async';
+import 'dart:isolate';
+
 import 'package:flutter/material.dart';
 import 'package:temp_app/constants/environment.dart';
 
@@ -6,5 +9,20 @@ import 'app.dart';
 void main() async {
   await Environment.init();
 
-  runApp(App());
+  FlutterError.onError = Environment.recordFlutterError;
+
+  Isolate.current.addErrorListener(RawReceivePort(
+    (pair) async {
+      final List<dynamic> errorAndStacktrace = pair;
+      await Environment.recordError(
+        errorAndStacktrace.first,
+        errorAndStacktrace.last,
+      );
+    },
+  ).sendPort);
+
+  runZonedGuarded(
+    () => runApp(App()),
+    Environment.recordError,
+  );
 }
